refactor: mutation 로직 통합
